refactor: move `EVMSliceExec` to its own file
